Creating a new multi-module workspace. - Removed delegator.go (Will recreate later) - added new worker path module. - moved worker related files to the worker path.
