+++ go.work
@@ -1,0 +1,3 @@
+go 1.21.4
+
+use ./worker
