+++ worker/go.mod
@@ -1,0 +1,5 @@
+module worker
+
+go 1.21.4
+
+require github.com/google/uuid v1.4.0
