+++ worker/writer.go
@@ -1,4 +1,4 @@
-package main
+package worker
 
 // Writer updates the seen nodes list, and sends found nodes downstream to the database.
 type writer struct{}
